moved widgets to /post instead of /thread, added basic post toolbar widget, added some colors to dark theme
