+++ lib/widgets/post/toolbar.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:knocky/models/thread.dart';
+
+class Toolbar extends StatelessWidget {
+  final ThreadPost post;
+
+  Toolbar({@required this.post});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        color: Theme.of(context).primaryColor,
+        height: 30,
+        child: Row() // todo: post metadata + controls
+        );
+  }
+}
